Handling R-URI if domain is present in from of @MSISDN of caller (seen in Samsung phones)
In Samsung Tab A when calling the R-URI is in following form

sip:0398765432100;phone-context=0498765432100@0498765432100;user=phone

which results in calling not being connected as the domain part after @ wont be resolved by S-CSCF.
Therefore, fix is to replace domain part with S-CSCF domain name.
This fix may not hold good in case of Roaming user I guess.